decouple CMake and compile_commands.json (#32)
Presence of a compilation database does not necessarily indicate a
project uses CMake (although disaster still requires that a CMake
project contain compile_commands.json). For example,
compile_commands.json is used by the clangd LSP and other tooling. If
a `CMakeLists.txt` is found in the project root, compile with `cmake`;
otherwise, compile with the command found in the compilation database.

`disaster-get-object-file-path-cmake` will now use the last object
file specified in the compile command, matching the behavior of
Clang/GCC.

Fixed a naive comparison in
`disaster-create-compile-command-compile-db` to work with paths that
are semantically equal, but not literally equal. Both
`file-name-directory` and `vc-root-dir` return directories with
trailing slashes, so relative file names need not begin with a slash;
however, the `"directory"` value in compile_commands.json does not
necessarily end in a slash. Extra slashes are harmless on most unix
systems, but they ought to be removed for aesthetic purposes.